This commit/PR does 2 things:
1. In the prior release I thought I didn't need the Java
Class.getResource() methods for obtaining resources.  Well, I was wrong.
So this reinstalls these important methods with a major improvement:
These methods will now work from JAR files of classes as well as from
within a file system.

2. I increased my detection of all kinds of minor code quality issues.
So this commit fixes all that I could find. Examples: redundant casts,
redundant declaration of generic parameters, methods that could be
declared static, variable name hiding, etc.

+++ src/test/java/org/apache/datasketches/kll/KllDoublesSketchTest.java
@@ -119,7 +119,7 @@ public class KllDoublesSketchTest {
       double rank = i/10.0;
       double q = rank == 1.0 ? i : i + 1;
       assertEquals(sketch.getQuantile(rank, EXCLUSIVE), q);
-      q = (double)(rank == 0 ? i + 1.0 : i);
+      q = rank == 0 ? i + 1.0 : i;
       assertEquals(sketch.getQuantile(rank, INCLUSIVE), q);
     }
